projects/lead-generation/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
twilio==8.10.0
